web/rules/requirements.txt
pyyaml==6.0.2
pygments==2.18.0
